Clean up of model and task
Update task
Remove utils file
Update args
Update README.md
Update input_fn to support from_tensor_slices
