+++ imdb/keras/trainer/model.py
@@ -17,7 +17,6 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from . import utils
 import numpy as np
 
 import tensorflow as tf
@@ -61,35 +60,31 @@ def keras_estimator(model_dir, config, learning_rate, vocab_size):
   return estimator
 
 
-def input_fn(x, y, batch_size, mode):
+def input_fn(features, labels, batch_size, mode):
   """Input function.
 
   Args:
-    x: (numpy.array) Training or eval data.
-    y: (numpy.array) Labels for training or eval data.
+    features: (numpy.array) Training or eval data.
+    labels: (numpy.array) Labels for training or eval data.
     batch_size: (int)
     mode: tf.estimator.ModeKeys mode
 
   Returns:
     A tf.estimator.
   """
-  # Default settings for training
-  num_epochs = None
-  shuffle = True
-
-  # Override if this is eval
+  # Default settings for training.
+  if labels is None:
+    inputs = features
+  else:
+    labels = np.asarray(labels).astype('int').reshape((-1, 1))
+    inputs = (features, labels)
+  # Convert the inputs to a Dataset.
+  dataset = tf.data.Dataset.from_tensor_slices(inputs)
+  if mode == tf.estimator.ModeKeys.TRAIN:
+    dataset = dataset.shuffle(1000).repeat().batch(batch_size)
   if mode == tf.estimator.ModeKeys.EVAL:
-    num_epochs = 1
-    shuffle = False
-  y = np.asarray(y).astype('int').reshape((-1, 1))
-
-  return tf.estimator.inputs.numpy_input_fn(
-      x,
-      y=y,
-      batch_size=batch_size,
-      num_epochs=num_epochs,
-      shuffle=shuffle,
-      queue_capacity=50000)
+    dataset = dataset.batch(batch_size)
+  return dataset.make_one_shot_iterator().get_next()
 
 
 def serving_input_fn():
@@ -106,45 +101,3 @@ def serving_input_fn():
                                                         feature_placeholder)
 
 
-def train_and_evaluate(output_dir, hparams):
-  """Helper function: Trains and evaluate model.
-
-  Args:
-    output_dir: (str) File path where training files will be written.
-    hparams: (dict) Command line parameters passed from task.py
-  """
-  # Load data.
-  (train_data, train_labels), (test_data, test_labels) = \
-      utils.prepare_data(train_data_file=hparams['train_file'],
-                         word_index_file=hparams['word_index_file'],
-                         num_words=TOP_K)
-  # Create estimator.
-  run_config = tf.estimator.RunConfig(save_checkpoints_steps=500)
-  estimator = keras_estimator(model_dir=output_dir,
-                              config=run_config,
-                              learning_rate=hparams['learning_rate'],
-                              vocab_size=VOCAB_SIZE)
-  train_steps = hparams['num_epochs'] * len(train_data) / hparams['batch_size']
-  # Create TrainSpec.
-  train_spec = tf.estimator.TrainSpec(
-      input_fn=input_fn(
-          train_data,
-          train_labels,
-          hparams['batch_size'],
-          mode=tf.estimator.ModeKeys.TRAIN),
-      max_steps=train_steps)
-  # Create EvalSpec.
-  exporter = tf.estimator.LatestExporter('exporter', serving_input_fn)
-  eval_spec = tf.estimator.EvalSpec(
-      input_fn=input_fn(
-          test_data,
-          test_labels,
-          hparams['batch_size'],
-          mode=tf.estimator.ModeKeys.EVAL),
-      steps=None,
-      exporters=exporter,
-      start_delay_secs=10,
-      throttle_secs=10)
-
-  # Start training
-  tf.estimator.train_and_evaluate(estimator, train_spec, eval_spec)
